fix: renamed model file
